Add vcf.to.numeric method
Previously, I had been using the as.numeric function to convert
allelic depth and genotype data from the VCF file's text into numbers
in R. However, the VCF format allows for the use of the period (.) to
indicate that no measurement exists. The as.numeric method will
convert this to an NA, but it throws a lot of warnings. The new
vcf.to.numeric function will not throw warnings and will properly
convert both numeric data and period characters as well as stop the
program if any other character is found.