Remove LabyrinthImputeHelper and improve runtime output
LabyrinthImputeHelper was an unnecessary function, and the code and
usage is cleaner without it. All of the printing that happens while
the code runs was cleaned up as well so that reading the VCF file,
running the imputation, and writing the results are all times
seperately and printed. The total runtime is also still calculated and
displayed.