Make correction to imputation algorithm
In the original LB-Impute, all sites where all three hidden
states (homozygous parent 1, homozygous parent 2, heterozygous) have
equal probability are not used for the viterbi algorithm and are added
back in afterward. We were using them all. This is being changed now
in order to make the algorithm more identical to LB-Impute for
testing, but it may be worth changing it back later.