Modify most functions to take probability vector instead of function
Most of the functions required a recombination function to be passed
as a parameter, but it is easier to require the user to pass a
probability vector since they would end up having to compute it
anyway. The first 5 tests were also adapted and retested, and
everything looks good so far.