Don't use multiple models
The central biological assumption to this point had been that when
crossover occurs, only one of the sister chromatids from each
homologous chromosome is involved. This required running imputation on
a number of different models. I have been informed, now, that that is
not correct, thus we can use only a single hidden Markov model instead
of many.