Remove deprecated parse_common_F1_probs_to_R function
This function was written at a time where I thought maybe a single F1 plant
was the common ancestor of all the other is an F{x} generation. This
is still possibly the case, but the code was never used, and will be
cleaned up later to generalize it even more.