Regenerate filtered and masked versions of 4 datasets (identical)
The filtered and masked datasets where regenerated for each of the 4
curent datasets using the `LabyrinthPreparePublicationData`
function. The resulting files are identical to the previous files, but
now they are easily reproducible. The file names of the results are
slightly different as indicated in the diff. A log of the results is
included as well.